api/requirements.txt
fastapi
uvicorn
pydantic
gliner
pymongo
python-dotenv
huggingface_hub
certifi
pip-system-certs
PyJWT
dotenv
sqlalchemy
psycopg2-binary
